Move internal/multierror to daemon/internal/multierror
