+++ daemon/containerd/cache.go
@@ -13,8 +13,8 @@ import (
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/image/cache"
 	"github.com/docker/docker/daemon/internal/layer"
+	"github.com/docker/docker/daemon/internal/multierror"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/multierror"
 	"github.com/moby/moby/api/types/backend"
 	"github.com/moby/moby/api/types/container"
 	"github.com/opencontainers/go-digest"
